pyproject.toml
[tool.poetry]
name = "mantra bot"
version = "0.1.0a"
description = "A Powerful and Easy to use Discord bot 🤖"
authors = ["Nishant Sapkota <snishant306@gmail.com>", "Stuti Upreti <stutiupreti12@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
pydantic = {extras = ["dotenv"], version = "^1.10.2"}
hikari = "^2.0.0.dev111"
hikari-lightbulb = "^2.2.4"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
aerich = "^0.7.1"
hikari-yuyo = "^1.1.1a1"
hikari-miru = {git = "https://github.com/HyperGH/hikari-miru", rev = "05ac0b4"}
aioredis = "^2.0.1"
asyncpraw = "^7.5.0"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.8.0", allow-prereleases = true}
pre-commit = "^2.20.0"


[tool.aerich]
tortoise_orm = "mantra.core.tortoise_config.tortoise_config"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
